pyproject.toml
[tool.poetry]
name = "eduhub"
version = "0.1.0"
description = ""
authors = ["sina rasouli <sinarasouli79@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.1"
python-dotenv = "^1.0.1"
psycopg = {extras = ["binary"], version = "^3.1.18"}
django-debug-toolbar = "^4.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
